+++ app/src/main/java/com/magma/tradecoach/di/LiveDataExtensions.kt
@@ -1,0 +1,9 @@
+package com.magma.tradecoach.di
+
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
+
+fun<T : Any, L : LiveData<T>> LifecycleOwner.observe(liveData: L, body: (T) -> Unit) {
+    liveData.observe(this, Observer(body))
+}
