+++ app/src/main/java/com/magma/tradecoach/utilities/Constants.kt
@@ -2,5 +2,7 @@ package com.magma.tradecoach.utilities
 
 object Constants {
       const val LOGGED_KEY = "STAY_LOGGED"
-
+      const val PURCHASE_ITEM_COUNT = 6
+      const val PURCHASE_GRID_COUNT = 3
+      const val CHAT_LIMIT = 50
 }
